fix: Load the commit template via husky
Git does not have an automated way to load the git template for
everyone (without the use of --local settings).  A common solution
is to use a plugin for husky that simply loads the template.
There were a few options out there, but they didn't work with the
latest version of husky, so we just wrote our own.  If someone is
doing a straight commit using the editor, it will use the template,
otherwise, the user is doing something like `--amend`, and we won't
mess with the message (because presumably, they've already seen the
template).

Semver: patch
Ref: LOG-7458