refactor: A value of empty string should apply the default value
When a value of '' is given, this should trigger the default
value to be used. A common use case is in k8s manifests where
mustache templates may blindly set an env var to a template
value. If that value is empty, the empty string will be used
and can cause the code to throw for types like `enum`. This
change allows us to skip setting defaults in such manifests
and let `Config` handle the defaults.

Also, the `_validateString` function was removed due to it
not being covered. The condition for `_required` is covered
in `_populate`, leaving the string validation function to
only be used for lists. This error check was simply moved
to the list validation.

In addition, a new mutex error has been added for using
`required` and `default` together, which doesn't make much
sense at a glance.

BREAKING CHANGE: This change affects the way that default values
are assigned, and the default value will now be used if the
env var's value is ''. Also, using `.required()` and `.default()`
in the same definition will now result in a mutex error.