feat: allowEmpty() will allow '' to be a valid value
In certain cases, we may want to have a default for use if
the env var is *omitted*, but allow the value of the empty
string to be a valid value (without triggering the default
value). This allows for that.

`.string('db-options).default('autoConnect=true').allowEmpty()`

would allow the db-options to be set to an empty string and
NOT auto connect in this example.