feat: add autocomplete and type inference
Adds autocomplete to `config.get` and type
inference to the value assigned. For example
if you type `config.get('` you'll see a list
of suggested values that come directly from
whatever was configured by the user in `config.js`.
If you've confgiured "port" to be a number in
`config.js` and assign a variable like so:
`const port = config.get('port')`
The variable "port" will be inferred to be a number.
Auto-complete works under JavaScript, REPL, and
TypeScript. JavaScript users can continue to use
the Config constructor method but must switch
to using the `createConfig` helper method to
enjoy auto-complete.

Ref: NA